requirements.txt
apprise==1.3.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
Markdown==3.4.3
oauthlib==3.2.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
soupsieve==2.4.1
urllib3==2.0.2
